+++ tests/csr/arch.rs
@@ -15,7 +15,7 @@ pub struct Core {
     pub mscratch: xlenbits,
     pub sscratch: xlenbits,
     pub cur_privilege: Privilege,
-    pub Xs: [xlenbits;32],
+    pub Xs: [xlenbits; (32 as usize)],
     pub config: Config,
 }
 
@@ -24,9 +24,9 @@ pub struct Config {
 
 }
 
-pub const xlen: usize = 64;
+pub const xlen: i128 = 64;
 
-pub const xlen_bytes: usize = 8;
+pub const xlen_bytes: i128 = 8;
 
 pub type xlenbits = BitVector<xlen>;
